feat: 메인페이지 설정 변경
